--- app/controllers/users/movies_controller.rb
@@ -1,17 +1,0 @@
-module Users
-  class MoviesController < ApplicationController
-    before_action :require_user
-
-    def index
-      if params[:top_rated]
-        @movies = MoviesFacade.top_rated
-      elsif params[:search]
-        @movies = MoviesFacade.search(params[:search])
-      end
-    end
-
-    def show
-      @movie = MoviesFacade.details(params[:movie_id])
-    end
-  end
-end
