+++ app/views/_user_dashboard.html.erb
@@ -14,7 +14,7 @@
     <% @user.parties(true).each do |party| %>
         <p><%= image_tag @image_url_hash[party.movie_id], size:"250x300", alt: "#{party.movie_title} Poster" %></p>
 
-        <p><%= link_to party.movie_title, "/dashboard/movies/#{party.movie_id}" %></p>
+        <p><%= link_to party.movie_title, movie_path(party.movie_id) %></p>
       <li>
         <h3>Partygoers:</h3>
         <% party.users.each do |user| %>
@@ -38,7 +38,7 @@
       <li>
         <p><%= image_tag @image_url_hash[party.movie_id], size:"250x300", alt: "#{party.movie_title} Poster" %></p>
 
-        <p><%= link_to party.movie_title, "/dashboard/movies/#{party.movie_id}" %></p>
+        <p><%= link_to party.movie_title, movie_path(party.movie_id) %></p>
         <h3>Partygoers:</h3>
 
         <% party.users.each do |user| %>
