+++ app/views/discover/index.html.erb
@@ -1,6 +1,6 @@
 <h1>User Discover Index</h1>
-<p><%= button_to 'Discover Top Rated Movies', dashboard_movies_path, method: :get, params: {top_rated: true} %></p>
-<%= form_with url: dashboard_movies_path, method: :get, local: true do |form| %>
+<p><%= button_to 'Discover Top Rated Movies', movies_path, method: :get, params: {top_rated: true} %></p>
+<%= form_with url: movies_path, method: :get, local: true do |form| %>
   <%= form.label :search, 'Search' %>
   <%= form.text_field :search %>
   <%= form.submit 'Search by Movie Title' %>
