+++ app/views/movies/show.html.erb
@@ -1,6 +1,6 @@
 <h1><%= @movie.title %></h1>
 
-<p><%= button_to "Create a Viewing Party for #{@movie.title}", "/dashboard/movies/#{@movie.id}/viewing_parties/new", method: :get %></p>
+<p><%= button_to "Create a Viewing Party for #{@movie.title}", new_movie_viewing_party_path(@movie.id), method: :get %></p>
 <p><%= button_to "Back to Discover Page", discover_path, method: :get %></p>
 
 <p>Vote Average: <%= @movie.vote_average %></p>
