+++ config/routes.rb
@@ -15,11 +15,13 @@ Rails.application.routes.draw do
 
   post '/movies_top_rated', to: 'movies#top_rated'
   post '/movies_search', to: 'movies#search'
-  get '//discover', to: 'discover#index'
-  get '/dashboard/movies', to: 'users/movies#index'
-  get '/dashboard/movies/:movie_id', to: 'users/movies#show'
+  get '/discover', to: 'discover#index'
 
-  get '/dashboard/movies/:movie_id/viewing_parties/new', to: 'users/viewing_parties#new'
-  post '/dashboard/movies/:movie_id/viewing_parties', to: 'users/viewing_parties#create'
-  get '/dashboard/movies/:movie_id/viewing_parties/:id', to: 'users/viewing_parties#show'
+  resources :movies, only: [:index, :show] do
+    resources :viewing_parties, only: [:new, :create, :show]
+  end
+
+  # get '/dashboard/movies/:movie_id/viewing_parties/new', to: 'users/viewing_parties#new'
+  # post '/dashboard/movies/:movie_id/viewing_parties', to: 'users/viewing_parties#create'
+  # get '/dashboard/movies/:movie_id/viewing_parties/:id', to: 'users/viewing_parties#show'
 end
