+++ spec/models/viewing_party_spec.rb
@@ -1,5 +1,3 @@
-# frozen_string_literal: true
-
 require 'rails_helper'
 
 RSpec.describe ViewingParty, type: :model do
